requirements.txt
streamlit>=1.24.0
mistralai>=0.0.7
transformers>=4.30.0
python-docx>=0.8.11
torch>=2.0.0
torchaudio>=2.0.0
requests>=2.31.0
tiktoken>=0.5.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.24.0
python-dotenv>=1.0.0